Install redux RTK and API routing
