+++ package.json
@@ -13,9 +13,11 @@
     "@emotion/styled": "^11.13.0",
     "@mui/icons-material": "^6.1.6",
     "@mui/material": "^6.1.6",
+    "@reduxjs/toolkit": "^2.3.0",
     "next": "15.0.2",
     "react": "19.0.0-rc-02c0e824-20241028",
-    "react-dom": "19.0.0-rc-02c0e824-20241028"
+    "react-dom": "19.0.0-rc-02c0e824-20241028",
+    "react-redux": "^9.1.2"
   },
   "devDependencies": {
     "eslint": "^8",
