+++ src/containers/ServiceListContainer/index.js
@@ -2,21 +2,15 @@ import { useState } from "react";
 import ServiceTable from "./ServiceTable";
 import NewService from "./NewService";
 import EditService from "./EditService";
-
-const MOCK_SERVICES = [
-  { id: 1, name: '半身調理', price: 500, period: 30 },
-  { id: 2, name: '全身調理', price: 1000, period: 60 },
-  { id: 3, name: '輕鬆調理', price: 300, period: 15 },
-  { id: 4, name: '足底按摩', price: 500, period: 30 },
-  { id: 5, name: '全身去角質', price: 1350, period: 60 },
-]
+import { useGetServicesQuery } from "@/features/api/apiSlice";
 
 export const PERIODS = [15, 30, 60, 90, 120];
 
 const ServiceListContainer = () => {
-  const [services, setServices] = useState(MOCK_SERVICES);
+  const { data, isLoading } = useGetServicesQuery();
+  const services = data?.services;
   const [editServiceId, setEditServiceId] = useState(null);
-  const editService = services.find(service => service.id === editServiceId);
+  const editService = editServiceId ? services?.find(service => service.id === editServiceId) : null;
 
   const handleSubmit = (service) => {
     setServices([...services, service]);
@@ -34,6 +28,8 @@ const ServiceListContainer = () => {
     setServices(services.map(s => s.id === service.id ? service : s));
     setEditServiceId(null);
   }
+
+  if (isLoading) return <div>Loading...</div>;
 
   return <>
     <NewService onSubmit={handleSubmit} />
